Muzix application with tasks

+++ src/main/java/com/stackroute/service/TrackServiceImpl.java
@@ -1,5 +1,7 @@
 package com.stackroute.service;
 import com.stackroute.domain.Track;
+import com.stackroute.exceptions.TrackAlreadyExistsException;
+import com.stackroute.exceptions.TrackNotFoundException;
 import com.stackroute.repository.TrackRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -10,16 +12,22 @@ import java.util.Optional;
 @Service
 public class TrackServiceImpl implements TrackService {
 
-    TrackRepository trackRepository;
+    public TrackRepository trackRepository;
     @Autowired
     public TrackServiceImpl(TrackRepository trackRepository) {
         this.trackRepository = trackRepository;
     }
 
     @Override
-    public Track saveTrack(Track track) {
-       Track savedTrack=trackRepository.save(track);
-       return savedTrack;
+    public Track saveTrack(Track track)throws  TrackAlreadyExistsException {
+        if(trackRepository.existsById(track.getTrackId())){
+            throw new TrackAlreadyExistsException("Track already exists");
+        }
+        else {
+            Track savedTrack = trackRepository.save(track);
+            return savedTrack;
+        }
+
     }
 
     @Override
@@ -28,9 +36,16 @@ public class TrackServiceImpl implements TrackService {
     }
 
     @Override
-    public  Optional<Track> getTrackById(int id) {
-       Optional<Track> getTrack= trackRepository.findById(id);
-       return getTrack;
+    public  Optional<Track> getTrackById(int id) throws TrackNotFoundException{
+        if(trackRepository.existsById(id)) {
+            Optional<Track> getTrack = trackRepository.findById(id);
+            return getTrack;
+        }
+        else
+        {
+            throw new TrackNotFoundException("Track not present");
+        }
+
     }
 
     @Override
@@ -40,8 +55,20 @@ public class TrackServiceImpl implements TrackService {
     }
 
     @Override
-    public List<Track> deleteTrackById(int id) {
-        trackRepository.deleteById(id);
-        return trackRepository.findAll();
+    public List<Track> deleteTrackById(int id) throws TrackNotFoundException {
+        if(trackRepository.existsById(id)) {
+            trackRepository.deleteById(id);
+            return trackRepository.findAll();
+        }
+        else
+        {
+            throw new TrackNotFoundException("Track Not found");
+        }
+    }
+
+    @Override
+    public List<Track> findByName(String trackName) throws TrackNotFoundException {
+            return trackRepository.findByName(trackName);
+
     }
 }
